fix(sdk): widen isolated uid/gid to long for uint32 range
C#/Kotlin int could not hold values above Int32.MaxValue; align with
execd-api uint32 and add deserialization/serialization regression tests.

+++ sdks/sandbox/kotlin/sandbox/src/main/kotlin/com/alibaba/opensandbox/sandbox/infrastructure/adapters/service/IsolatedSessionsAdapter.kt
@@ -55,8 +55,8 @@ private data class IsolatedCreateBody(
     val binds: List<BindMountBody>? = null,
     val share_net: Boolean? = null,
     val env_passthrough: EnvPassthroughBody? = null,
-    val uid: Int? = null,
-    val gid: Int? = null,
+    val uid: Long? = null,
+    val gid: Long? = null,
     val uid_mode: String? = null,
     val idle_timeout_seconds: Int? = null,
 )
